feat: Implement comprehensive fix for infinite recursion
Implements a 5-phase plan to address the infinite re-rendering issue caused by cascading circular synchronization between SessionContext and SimpleOrganizationProvider.

Phase 1: Establishes SessionContext as the single source of truth for organization state, simplifying SimpleOrganizationProvider to a read-only view.
Phase 2: Fixes dependency stability by applying memoization to hooks and callbacks.
Phase 3: Eliminates competing organization selection logic by centralizing it within SessionContext.
Phase 4: Introduces safeguards such as debouncing and duplicate event prevention.
Phase 5: Optimizes for multi-organization users with caching and batching.
